refactor(workflow): relocate triage parser to app/workflow package
