+++ app/workflow/issue_parser.go
@@ -1,4 +1,4 @@
-package bot
+package workflow
 
 import (
 	"encoding/json"
