+++ app/workflow/issue_parser_test.go
@@ -1,4 +1,4 @@
-package bot
+package workflow
 
 import (
 	"strings"
